% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-replication.R
\name{arp}
\alias{arp}
\title{Aggregated replication probability}
\usage{
arp(R, pi, power, alpha)
}
\arguments{
\item{R}{The number of studies considering the original study. For one-to-one design R = 2}

\item{pi}{The theory strenght i.e. the proportion of true effects}

\item{power}{The power of the R studies. Power can be a single value (used for R studies) or a vector of length R}

\item{alpha}{The alpha level. For the replication studies (R - 1) alpha/2 will be used}
}
\value{
The aggregated probability
}
\description{
Aggregated replication probability
}
\examples{
arp(2, 0.2, 0.8, 0.05) # as reported in Miller (2009) p.619
}
