% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-replication.R
\name{bf_replication}
\alias{bf_replication}
\title{bf_replication}
\usage{
bf_replication(mu_original, se_original, replication)
}
\arguments{
\item{mu_original}{}

\item{se_original}{}

\item{replication}{}
}
\description{
bf_replication
}
