% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-replication.R
\name{small_telescope}
\alias{small_telescope}
\title{small_telescope}
\usage{
small_telescope(or_d, or_se, rep_d, rep_se, small, ci = 0.95)
}
\arguments{
\item{or_d}{}

\item{or_se}{}

\item{rep_d}{}

\item{rep_se}{}

\item{small}{}

\item{ci}{}
}
\description{
small_telescope
}
